+++ pkg/http/lib/src/exception.dart
@@ -1,0 +1,17 @@
+// Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library http.exception;
+
+/// An exception caused by an error in a pkg/http client.
+class ClientException implements Exception {
+  final String message;
+
+  /// The URL of the HTTP request or response that failed.
+  final Uri uri;
+
+  ClientException(this.message, [this.uri]);
+
+  String toString() => message;
+}