+++ pkg/oauth2/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.9.3
+
+* Update the `http` dependency.
+
 # 0.9.2
 
 * Expand the dependency on the HTTP package to include 0.10.x.